Fix check 7

+++ 0x15-file_io/0-read_textfile.c
@@ -8,7 +8,7 @@
  */
 ssize_t read_textfile(const char *filename, size_t letters)
 {
-	int rc, file, c;
+	ssize_t rc, file, c;
 	char *data;
 
 	data = malloc(sizeof(char) * letters);
